refactor: api 폴더에서 api 관련 코드를 관리하도록 변경
