+++ src/main/java/com/ziine/admin/artwork/application/AdminArtworkReviewService.java
@@ -6,10 +6,10 @@ import com.ziine.admin.artwork.domain.repository.ArtworkStatusHistoryRepository;
 import com.ziine.admin.artwork.dto.request.AdminArtworkRejectRequestDto;
 import com.ziine.admin.auth.application.AdminContextHolder;
 import com.ziine.admin.auth.domain.Admin;
-import com.ziine.domains.artwork.application.exception.ArtworkNotFoundException;
-import com.ziine.domains.artwork.domain.entity.ArtworkEntity;
-import com.ziine.domains.artwork.domain.entity.ArtworkStatus;
-import com.ziine.domains.artwork.domain.repository.ArtworkRepository;
+import com.ziine.api.artwork.application.exception.ArtworkNotFoundException;
+import com.ziine.api.artwork.domain.entity.ArtworkEntity;
+import com.ziine.api.artwork.domain.entity.ArtworkStatus;
+import com.ziine.api.artwork.domain.repository.ArtworkRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
