+++ src/main/java/com/ziine/admin/artwork/domain/entity/ArtworkStatusHistoryEntity.java
@@ -1,7 +1,7 @@
 package com.ziine.admin.artwork.domain.entity;
 
-import com.ziine.domains.artwork.domain.entity.ArtworkEntity;
-import com.ziine.domains.artwork.domain.entity.ArtworkStatus;
+import com.ziine.api.artwork.domain.entity.ArtworkEntity;
+import com.ziine.api.artwork.domain.entity.ArtworkStatus;
 import jakarta.persistence.Column;
 import jakarta.persistence.ConstraintMode;
 import jakarta.persistence.Entity;
