+++ src/main/java/com/ziine/admin/artwork/presentation/AdminArtworkRetrieveController.java
@@ -2,7 +2,7 @@ package com.ziine.admin.artwork.presentation;
 
 import com.ziine.admin.artwork.application.AdminArtworkRetrieveService;
 import com.ziine.admin.artwork.dto.response.AdminArtworkRetrieveResponseDto;
-import com.ziine.domains.artwork.domain.entity.ArtworkStatus;
+import com.ziine.api.artwork.domain.entity.ArtworkStatus;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
