+++ src/main/java/com/ziine/api/artist/domain/entity/ContactEntity.java
@@ -1,0 +1,59 @@
+package com.ziine.api.artist.domain.entity;
+
+import com.ziine.api.artwork.dto.request.ArtworkPersistRequestDto.ContactRequestDto;
+import jakarta.persistence.Column;
+import jakarta.persistence.ConstraintMode;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.ForeignKey;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+@Table(name = "contact")
+public class ContactEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "contact_id")
+    private Long id;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false, length = 20)
+    private ContactType type;
+
+    @Column(nullable = false, length = 255)
+    private String value;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "artist_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT), nullable = false)
+    private ArtistEntity artistEntity;
+
+    public ContactEntity(
+        final ContactType type,
+        final String value,
+        final ArtistEntity artistEntity
+    ) {
+        this.type = type;
+        this.value = value;
+        this.artistEntity = artistEntity;
+    }
+
+    public static ContactEntity fromContactRequestDto(
+        final ContactRequestDto contactRequestDto,
+        final ArtistEntity artistEntity
+    ) {
+        return new ContactEntity(contactRequestDto.type(), contactRequestDto.value(), artistEntity);
+    }
+}
