+++ src/main/java/com/ziine/api/artist/domain/entity/ContactType.java
@@ -1,0 +1,17 @@
+package com.ziine.api.artist.domain.entity;
+
+import lombok.Getter;
+
+@Getter
+public enum ContactType {
+
+    INSTAGRAM("인스타그램"),
+    WEBSITE("웹사이트"),
+    OTHER("기타");
+
+    private final String description;
+
+    ContactType(final String description) {
+        this.description = description;
+    }
+}
