+++ src/main/java/com/ziine/api/artist/domain/entity/EducationEntity.java
@@ -1,0 +1,43 @@
+package com.ziine.api.artist.domain.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.ConstraintMode;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.ForeignKey;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Entity
+@Table(name = "education")
+public class EducationEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "education_id")
+    private Long id;
+
+    @Column(nullable = false, length = 20)
+    private String tag;
+
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    @JoinColumn(name = "artist_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT), nullable = false)
+    private ArtistEntity artistEntity;
+
+    public EducationEntity(
+        final String tag,
+        final ArtistEntity artistEntity
+    ) {
+        this.tag = tag;
+        this.artistEntity = artistEntity;
+    }
+}
