+++ src/main/java/com/ziine/api/artist/domain/repository/ArtistRepository.java
@@ -1,0 +1,10 @@
+package com.ziine.api.artist.domain.repository;
+
+import com.ziine.api.artist.domain.entity.ArtistEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ArtistRepository extends JpaRepository<ArtistEntity, Long> {
+
+}
