+++ src/main/java/com/ziine/api/artwork/application/ArtworkService.java
@@ -1,0 +1,25 @@
+package com.ziine.api.artwork.application;
+
+import com.ziine.api.artist.domain.entity.ArtistEntity;
+import com.ziine.api.artwork.domain.entity.ArtworkEntity;
+import com.ziine.api.artwork.domain.repository.ArtworkRepository;
+import com.ziine.api.artwork.dto.ArtworkPersistDto;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+
+@RequiredArgsConstructor
+@Service
+public class ArtworkService {
+
+    private final ArtworkRepository artworkRepository;
+
+    @Transactional
+    public ArtworkEntity persistArtwork(
+        final ArtworkPersistDto artworkPersistDto,
+        final ArtistEntity artistEntity
+    ) {
+        return artworkRepository.save(ArtworkEntity.fromArtworkPersistDto(artworkPersistDto, artistEntity));
+    }
+}
