--- src/main/java/com/ziine/domains/artist/domain/entity/ContactType.java
@@ -1,17 +1,0 @@
-package com.ziine.domains.artist.domain.entity;
-
-import lombok.Getter;
-
-@Getter
-public enum ContactType {
-
-    INSTAGRAM("인스타그램"),
-    WEBSITE("웹사이트"),
-    OTHER("기타");
-
-    private final String description;
-
-    ContactType(final String description) {
-        this.description = description;
-    }
-}
