--- src/main/java/com/ziine/domains/artist/domain/entity/ExhibitionEntity.java
@@ -1,57 +1,0 @@
-package com.ziine.domains.artist.domain.entity;
-
-import com.ziine.domains.artwork.dto.request.ArtworkPersistRequestDto.ExhibitionRequestDto;
-import jakarta.persistence.Column;
-import jakarta.persistence.ConstraintMode;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.ForeignKey;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.time.LocalDate;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "exhibition")
-public class ExhibitionEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "exhibition_id")
-    private Long id;
-
-    @Column(nullable = false, length = 80)
-    private String title;
-
-    @Column(nullable = false)
-    private LocalDate exhibitionDate;
-
-    @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    @JoinColumn(name = "artist_id", foreignKey = @ForeignKey(ConstraintMode.NO_CONSTRAINT), nullable = false)
-    private ArtistEntity artistEntity;
-
-    public ExhibitionEntity(
-        final String title,
-        final LocalDate exhibitionDate,
-        final ArtistEntity artistEntity
-    ) {
-        this.title = title;
-        this.exhibitionDate = exhibitionDate;
-        this.artistEntity = artistEntity;
-    }
-
-    public static ExhibitionEntity fromExhibitionRequestDto(
-        final ExhibitionRequestDto exhibitionRequestDto,
-        final ArtistEntity artistEntity
-    ) {
-        return new ExhibitionEntity(exhibitionRequestDto.title(), exhibitionRequestDto.exhibitionDate(), artistEntity);
-    }
-}
