--- src/main/java/com/ziine/domains/artist/domain/repository/ArtistRepository.java
@@ -1,10 +1,0 @@
-package com.ziine.domains.artist.domain.repository;
-
-import com.ziine.domains.artist.domain.entity.ArtistEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ArtistRepository extends JpaRepository<ArtistEntity, Long> {
-
-}
