--- src/main/java/com/ziine/domains/artwork/application/exception/ArtworkNotFoundException.java
@@ -1,13 +1,0 @@
-package com.ziine.domains.artwork.application.exception;
-
-import com.ziine.global.error.application.exception.BusinessException;
-import com.ziine.global.error.domain.ErrorCode;
-
-public class ArtworkNotFoundException extends BusinessException {
-
-    public static final ArtworkNotFoundException INSTANCE = new ArtworkNotFoundException();
-
-    private ArtworkNotFoundException() {
-        super(ErrorCode.ARTWORK_NOT_FOUND);
-    }
-}
