--- src/main/java/com/ziine/domains/artwork/dto/request/ArtworkPersistRequestDto.java
@@ -1,48 +1,0 @@
-package com.ziine.domains.artwork.dto.request;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.ziine.domains.artist.domain.entity.ContactType;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.PositiveOrZero;
-import java.time.LocalDate;
-import java.util.List;
-
-public record ArtworkPersistRequestDto(
-    @NotBlank(message = "Title is required")
-    String title,
-    @PositiveOrZero(message = "Width must be zero or greater")
-    int width,
-    @PositiveOrZero(message = "Height must be zero or greater")
-    int height,
-    @NotBlank(message = "Material is required")
-    String material,
-    @NotBlank(message = "Artwork image URL is required")
-    String artworkImageUrl,
-    @NotBlank(message = "Artist name is required")
-    String artistName,
-    String description,
-    List<String> educations,
-    List<ExhibitionRequestDto> exhibitions,
-    List<ContactRequestDto> contacts,
-    String email
-) {
-
-    public record ExhibitionRequestDto(
-        @NotBlank(message = "Exhibition title is required")
-        String title,
-        @JsonFormat(pattern = "yyyy-MM-dd")
-        LocalDate exhibitionDate
-    ) {
-
-    }
-
-    public record ContactRequestDto(
-        @NotNull(message = "Contact type is required")
-        ContactType type,
-        @NotBlank(message = "Contact value is required")
-        String value
-    ) {
-
-    }
-}
